Decorators e app modularizado.
