+++ backend/src/@winterFramework/app/winter.app.ts
@@ -1,0 +1,1 @@
+export * from './winterRunApplication'
