+++ backend/src/@winterFramework/decorator/winter.decorators.ts
@@ -1,0 +1,12 @@
+/*
+  Aqui é um barril contendo todos os decorators do WinterFramework.
+  NÃO IMPORTE 
+*/
+
+export * from './rest-controller.decorator'
+export * from './http-method.decorator'
+export * from './path-param.decorator'
+export * from './query-param.decorator'
+export * from './request-body.decorator'
+export * from './use-middleware.decorator'
+export * from './validate.decorator'
