--- backend/src/@winterFramework/index.ts
@@ -1,4 +1,0 @@
-// index.ts no diretório @winterFramework
-export * from './registry/controllerRegistry'
-export * from './decorator/decorators'
-export * from './app/winterRunApplication'
