+++ backend/src/@winterFramework/registry/controller.registry.ts
@@ -4,7 +4,6 @@ export const controllers: Map<any, { basePath: string; routes: any[] }> =
   new Map()
 export const paramMetadata: Map<any, Map<string, any[]>> = new Map()
 
-// Função para obter os controladores registrados
 export function getRegisteredControllers() {
   return [...controllersList]
 }
